chemie/requirements.txt
apipkg==1.4
asn1crypto==0.23.0
attrs==17.3.0
bleach==1.5.0
cffi==1.11.2
click==6.7
coverage==4.3.1
cryptography==2.1.4
Django==1.11
django-autocomplete-light==3.2.1
django-ckeditor==5.4.0
django-classy-tags==0.8.0
django-crontab==0.7.1
django-extended-choices==1.1.1
django-extensions==1.9.8
django-hashers-passlib==0.3
django-js-asset==0.1.1
django-markdown2==0.3.0
django-material==1.1.0
django-model-utils==3.0.0
django-mptt==0.8.7
django-nyt==1.0b5
django-post-office==2.0.8
django-recaptcha==1.2.0
django-sekizai==0.10.0
django-smart-selects==1.3.2
djangorestframework==3.5.3
execnet==1.5.0
factory-boy==2.9.2
Faker==0.8.4
first==2.0.1
freezegun==0.3.9
future==0.16.0
html5lib==0.9999999
idna==2.6
jsonfield==2.0.2
Markdown==2.6.9
markdown2==2.3.5
olefile==0.44
passlib==1.7.1
Pillow==4.3.0
pip-tools==1.10.1
pluggy==0.6.0
psycopg2==2.6.2
py==1.5.2
pycparser==2.18
pyOpenSSL==17.5.0
pytest==3.3.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-xdist==1.20.1
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
sorl-thumbnail==12.3
typing==3.6.2
uWSGI==2.0.15
Werkzeug==0.13
wiki==0.2
